Moved methods for demo setup from ResourceService to InitTestData
