+++ auth-server/src/main/java/com/antonio/authserver/service/ResourceService.java
@@ -1,36 +1,32 @@
 package com.antonio.authserver.service;
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
+import javax.transaction.Transactional;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Service;
 import com.antonio.authserver.dto.ResourceDto;
-import com.antonio.authserver.entity.Privilege;
 import com.antonio.authserver.entity.Resource;
 import com.antonio.authserver.entity.Role;
-import com.antonio.authserver.entity.RoleResourcePrivilege;
 import com.antonio.authserver.mapper.ResourceMapper;
 import com.antonio.authserver.model.CustomException;
 import com.antonio.authserver.repository.PrivilegeRepository;
 import com.antonio.authserver.repository.ResourceRepository;
 import com.antonio.authserver.repository.RoleRepository;
 import com.antonio.authserver.repository.RoleResourcePrivilegeRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Service;
-
-import javax.transaction.Transactional;
-import java.util.*;
 @Service
 @Transactional
 public class ResourceService {
     private final ResourceRepository resourceRepository;
     private final ResourceMapper resourceMapper;
-    private final PrivilegeRepository privilegeRepository;
     private final RoleRepository roleRepository;
-    private final RoleResourcePrivilegeRepository roleResourcePrivilegeRepository;
     @Autowired
-    public ResourceService(ResourceRepository resourceRepository, ResourceMapper resourceMapper, PrivilegeRepository privilegeRepository, RoleRepository roleRepository, RoleResourcePrivilegeRepository roleResourcePrivilegeRepository) {
+    public ResourceService(ResourceRepository resourceRepository, ResourceMapper resourceMapper, RoleRepository roleRepository) {
         this.resourceRepository = resourceRepository;
         this.resourceMapper = resourceMapper;
-        this.privilegeRepository = privilegeRepository;
         this.roleRepository = roleRepository;
-        this.roleResourcePrivilegeRepository = roleResourcePrivilegeRepository;
     }
 
     public List<ResourceDto> getAllResourcesFromDb(){
@@ -45,23 +41,6 @@ public class ResourceService {
         Resource resource = getResourceByNameOrThrowExceptionIfNotFound(oldName);
         resource.setName(newName);
         resourceRepository.save(resource);
-    }
-    //for basic role_user with all privileges
-    public Set<Resource> getBasicResourcesForRoleUserDemo() {
-        Set<Resource> resources = new HashSet<>();
-        resources.add(resourceRepository.findByName("COMPANIES").get());
-        resources.add(resourceRepository.findByName("BOOKS").get());
-        return resources;
-    }
-    //for basic role_user with all privileges
-    public void assignAllPrivilegesForRoleUser(Role role) {
-        Set<Resource> basicResources = getBasicResourcesForRoleUserDemo();
-        List<Privilege> privileges = privilegeRepository.findAll();
-        for (Resource resource : basicResources){
-            RoleResourcePrivilege roleResourcePrivilege = getRoleResourcePrivilegeByRoleAndResource(role, resource).get();
-            roleResourcePrivilege.getPrivileges().addAll(privileges);
-            roleResourcePrivilegeRepository.save(roleResourcePrivilege);
-        }
     }
 
     public void addResourceToDb(ResourceDto resourceDto){
@@ -95,8 +74,5 @@ public class ResourceService {
     }
     private Role getRoleByRealmNameAndNameOrThrowExceptionIfNotFound(String realmName,String roleName){
         return roleRepository.findByNameAndRealmName(roleName,realmName).orElseThrow(() -> new CustomException("Role with the name [" + roleName +"] could not be found!",HttpStatus.NOT_FOUND));
-    }
-    private Optional<RoleResourcePrivilege> getRoleResourcePrivilegeByRoleAndResource(Role role,Resource resource){
-        return roleResourcePrivilegeRepository.findByRoleAndResource(role, resource);
     }
 }
